refactor(server/events): extract makeCtrlChatObserver to observers/ctrl-chat.ts
Move chat-message CTRL_ROOM observer into a standalone factory. selectionBuffer
passed by reference (same Map instance as makeAwarenessObserver) so awareness
→ chat attachment path stays intact. Remove now-unused Y_MAP_CHAT, ChatMessage,
FlatOffset, validateRange, generateEventId imports from queue.ts.
